requirements-test.txt
beautifulsoup4==4.3.2
django-mptt==0.6.1
django-nose==1.3
flake8==2.3.0
freezegun==0.2.2
gnureadline==6.3.3
ipdb==0.8
ipython==2.4.1
jedi==0.8.1
mccabe==0.3
mock==1.0.1
nose==1.3.4
pep8==1.6.2
pyflakes==0.8.1
python-dateutil==2.3
